Minor change of Symfc.estimate_basis_set

+++ src/symfc/basis_sets/basis_sets_O3.py
@@ -256,7 +256,7 @@ class FCBasisSetO3(FCBasisSetBase):
         if self._fc_cutoff is None:
             n_sym, N = self._spg_reps._permutations.shape
             basis_size_estimates = 27 * (N**3) / n_sym / 6
-            return np.round(basis_size_estimates).astype(int)
+            return int(np.round(basis_size_estimates).astype(int))
 
         trans_perms = self._spg_reps.translation_permutations
         c_pt = compr_permutation_lat_trans_O3(
@@ -267,4 +267,4 @@ class FCBasisSetO3(FCBasisSetBase):
         )
         n_sym_prim = len(self._spg_reps._unique_rotations)
         basis_size_estimates = c_pt.shape[1] / n_sym_prim
-        return np.round(basis_size_estimates).astype(int)
+        return int(np.round(basis_size_estimates).astype(int))
